1. 思维控制器中,关于cmv的抵消和合并;   2. cmvAlgsModel去掉targetType;   3. 测试和bugfix;

+++ SMG_NothingIsAll/AIFoundation/AIAlg/AIImvAlgs.h
@@ -14,6 +14,6 @@
  *  MARK:--------------------输入mindValue--------------------
  *  所有值域,转换为0-10;(例如:hunger时0为不饿,10为非常饿)
  */
-+(void) commitIMV:(MVType)type from:(NSInteger)from to:(NSInteger)to;
++(void) commitIMV:(MVType)type from:(CGFloat)from to:(CGFloat)to;
 
 @end
