+++ SMG_NothingIsAll/AIFoundation/AIReactorControl/AIReactorControl.h
@@ -26,7 +26,7 @@
 -(void) createReactor:(AIMoodType)moodType;
 
 -(void) commitInput:(id)input;
--(void) commitIMV:(MVType)type from:(NSInteger)from to:(NSInteger)to;
+-(void) commitIMV:(MVType)type from:(CGFloat)from to:(CGFloat)to;
 -(void) commitCustom:(CustomInputType)type value:(NSInteger)value;
 
 @end
